Organize tests into subdirectories for controllers and other stuff
