+++ package.json
@@ -4,7 +4,7 @@
   "private": true,
   "scripts": {
     "start": "node ./bin/www",
-    "test": "NODE_ENV=test mocha",
+    "test": "NODE_ENV=test mocha --recursive",
     "docs": "jsdoc -c ./jsdoc.json -r",
     "coverage": "istanbul cover _mocha",
     "exec": "exec",
