+++ test/controllers/annotation_status_test.js
@@ -3,10 +3,10 @@
 const chai = require('chai');
 chai.use(require('chai-http'));
 
-const server = require('../app');
-const knex    = require('../app/lib/bookshelf').knex;
+const server = require('../../app/index');
+const knex    = require('../../app/lib/bookshelf').knex;
 
-const testdata = require('../seeds/test_data.json');
+const testdata = require('../../seeds/test_data.json');
 
 
 describe('Annotation Status Controller', function() {
