+++ test/controllers/gene_test.js
@@ -3,7 +3,7 @@
 const chai = require('chai');
 chai.use(require('chai-http'));
 
-const server = require('../app');
+const server = require('../../app/index');
 
 
 describe('Gene Controller', function() {
