+++ test/lib/publication_id_validator_test.js
@@ -2,7 +2,7 @@
 
 const chai = require('chai');
 
-const validator = require('../app/lib/publication_id_validator');
+const validator = require('../../app/lib/publication_id_validator');
 
 describe('Publication ID Validator', function() {
