+++ test/lib/uniprot_api_test.js
@@ -2,7 +2,7 @@
 
 const chai = require('chai');
 
-const Uniprot = require('../app/lib/uniprot_api');
+const Uniprot = require('../../app/lib/uniprot_api');
 
 describe('Uniprot API', function () {
